requirements.txt
requests>=2.32.3
numpy>=2.1.2
pillow>=10.4.0

